% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{berkeley}
\alias{berkeley}
\title{Berkeley Admissions dataset (Bickel et. al., 1975).}
\format{
A data frame with 4,526 rows and 3 variables:
\describe{
\item{gender}{gender of the individual}
\item{dept}{department of application, taking 6 values A to F}
\item{admit}{admission decision, taking two values Admitted/Rejected}
}
}
\usage{
berkeley
}
\description{
A dataset on admission statistics at UC Berkeley. Students apply to one of six
departments and the admission decision is based on gender and the department of choice.
}
\keyword{datasets}
