\name{cfa-internal}
\alias{bin.coeff}
\alias{bin.test}
\alias{callcfa}
\alias{perm.cfa}   
\alias{print.cfa}
\alias{print.hcfa}
\alias{print.mcfa}
\alias{shortcfa}
\alias{z.test}
\title{Internal cfa functions}
\description{
  Internal cfa functions.
}
\usage{
bin.coeff(n,k)
bin.test(ni,ep,ntotal)   
callcfa(configmatrix,cntvector)     
perm.cfa   
print.cfa(cfa.class.object) 
print.hcfa(cfa.class.object) 
print.mcfa(cfa.class.object) 
shortcfa(configmatrix,cntvector)    
z.test(ni,ep,ntotal)     
data(cfa2dat)   
}
\details{
  These are not to be called by the user.
}
\author{ Stefan Funke <funke@attglobal.net>}
\keyword{misc}
