% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex-compare.R
\name{check_constraints}
\alias{check_constraints}
\title{\itemize{
\item TeX(paste(phi_poly, "=", theta_poly)) *
}}
\usage{
check_constraints(x, mthd = "faircause_ranger")
}
\description{
\itemize{
\item TeX(paste(phi_poly, "=", theta_poly)) *
}
}
