% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompositions.R
\name{fairness_cookbook}
\alias{fairness_cookbook}
\title{fairness_cookbook}
\usage{
fairness_cookbook(data, X, W, Z, Y, x0, x1, nboot = 100, model = "ranger", ...)
}
\arguments{
\item{data}{Object of class \code{data.frame} containing the dataset.}

\item{X}{A \code{character} scalar giving the name of the
protected attribute. Must be one of the entries of \code{names(data)}.}

\item{W}{A \code{character} vector giving the names of all mediators.}

\item{Z}{A \code{character} vector giving the names of all confounders.}

\item{Y}{A \code{character} scalar giving the name of the outcome.}

\item{x0, x1}{Scalar values giving the two levels of the binary protected
attribute.}

\item{nboot}{An \code{integer} scalar determining the number of bootstrap
repetitions used when constructing confidence intervals of the estimates.}

\item{model}{A \code{character} scalar determining the model used in
estimation. Current options are \code{"ranger"} (default) which uses random
forests and \code{"linear"} which uses (generalized) linear models.}
}
\value{
An object of class \code{faircause}, containing estimates of the
causal fairness measures, together with some meta information.
}
\description{
Implementation of Fairness Cookbook in Causal Fairness Analysis
(Plecko & Bareinboim 2022). Uses only plain \code{R}.
}
\details{
The procedure takes the training and testing data as an input, together with
the causal graph given by an adjacency matrix and the list of resolving
variables, which should be kept fixed during the adaptation procedure. The
procedure then calculates a fair representation of the data, after which
any classification method can be used. There are, however, several valid
training options yielding fair predictions, and the best of them can be
chosen with cross-validation. For more details we refer the user to the
original paper. Most of the running time is due to the quantile regression
step using the ranger package.
}
\examples{
data <- faircause:::get_data_berkeley()

FCB <- fairness_cookbook(data, X = "gender", W = "dept", Z = character(0L),
                         Y = "admit", x0 = "Male", x1 = "Female")
FCB

}
\references{
Plecko, D. & Meinshausen, N. (2019).
Fair Data Adaptation with Quantile Preservation \cr
}
\author{
Drago Plecko
}
