% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompositions.R
\name{fairness_cookbook_eo}
\alias{fairness_cookbook_eo}
\title{fairness_cookbook_eo}
\usage{
fairness_cookbook_eo(data, X, W, Z, Y, Yhat, x0, x1, ylvl, ...)
}
\arguments{
\item{data}{Object of class \code{data.frame} containing the dataset.}

\item{X}{A \code{character} scalar giving the name of the
protected attribute. Must be one of the entries of \code{names(data)}.}

\item{W}{A \code{character} vector giving the names of all mediators.}

\item{Z}{A \code{character} vector giving the names of all confounders.}

\item{Y}{A \code{character} scalar giving the name of the outcome.}

\item{Yhat}{A \code{character} scalar giving the name of the constructed
predictor.}

\item{x0, x1}{Scalar values giving the two levels of the binary protected
attribute.}

\item{ylvl}{A value indicating within which group the decomposition should be
performed. For example, setting \code{ylvl = 1} would correspond to
performing causal effect decompositions for individuals who have
true outcome \code{Y == 1}.}

\item{...}{Further arguments passed to downstream model fitting functions.}
}
\description{
Implementation of Fairness Cookbook in Causal Fairness Analysis
(Plecko & Bareinboim 2022), for the case when equality of odds is the
measure of interest.
}
\details{
The procedure takes the data as an input, together with
the causal graph given by the Standard Fairness Model, and outputs a causal
decomposition of the EO measure into direct, indirect, and spurious effects.
}
