% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfm-proj.R
\name{SFM_proj}
\alias{SFM_proj}
\title{SFM_proj}
\usage{
SFM_proj(dataset = c("berkeley", "compas", "census"))
}
\arguments{
\item{dataset}{One of the following dataset names: \code{"berkeley"}, \code{"compas"},
or \code{"census"}}
}
\description{
Returns the Standard Fairness Model projection for the datasets that are
included in the faircause R-package.
}
