% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompositions.R
\name{fairness_cookbook}
\alias{fairness_cookbook}
\title{fairness_cookbook}
\usage{
fairness_cookbook(
  data,
  X,
  W,
  Z,
  Y,
  x0,
  x1,
  eo = FALSE,
  nboot = 100,
  model = "ranger",
  ...
)
}
\arguments{
\item{data}{Object of class \code{data.frame} containing the dataset.}

\item{X}{A \code{character} scalar giving the name of the
protected attribute. Must be one of the entries of \code{names(data)}.}

\item{W}{A \code{character} vector giving the names of all mediators.}

\item{Z}{A \code{character} vector giving the names of all confounders.}

\item{Y}{A \code{character} scalar giving the name of the outcome.}

\item{x0, x1}{Scalar values giving the two levels of the binary protected
attribute.}

\item{eo}{A \code{logical(1L)} indicating whether the object is of type
equality of odds or total variation. Default is false.}

\item{nboot}{An \code{integer} scalar determining the number of bootstrap
repetitions used when constructing confidence intervals of the estimates.}

\item{model}{A \code{character} scalar determining the model used in
estimation. Current options are \code{"ranger"} (default) which uses random
forests and \code{"linear"} which uses (generalized) linear models.}

\item{...}{Further arguments passed to downstream model fitting functions.}
}
\value{
An object of class \code{faircause}, containing estimates of the
causal fairness measures, together with some meta information.
}
\description{
Implementation of Fairness Cookbook in Causal Fairness Analysis
(Plecko & Bareinboim 2022). Uses only plain \code{R}.
}
\details{
The procedure takes the data as an input, together with
the causal graph given by the Standard Fairness Model, and outputs a causal
decomposition of the TV measure into direct, indirect, and spurious effects.
}
\examples{
\dontrun{
data <- faircause:::get_data_berkeley()

FCB <- fairness_cookbook(data, X = "gender", W = "dept", Z = character(0L),
                         Y = "admit", x0 = "Male", x1 = "Female")
FCB
}


}
\references{
Plecko, D. & Bareinboim, E. (2022).
Causal Fairness Analysis \cr
}
\author{
Drago Plecko
}
