% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompositions.R
\name{fairness_cookbook_eo}
\alias{fairness_cookbook_eo}
\title{fairness_cookbook_eo}
\usage{
fairness_cookbook_eo(
  data,
  X,
  Z,
  W,
  Y,
  Yhat,
  x0,
  x1,
  ylvl,
  method = c("medDML", "causal_forest"),
  model = c("ranger", "linear"),
  tune_params = FALSE,
  nboot1 = 1L,
  nboot2 = 100L,
  ...
)
}
\arguments{
\item{data}{Object of class \code{data.frame} containing the dataset.}

\item{X}{A \code{character} scalar giving the name of the
protected attribute. Must be one of the entries of \code{names(data)}.}

\item{Z}{A \code{character} vector giving the names of all mediators.}

\item{W}{A \code{character} vector giving the names of all confounders.}

\item{Y}{A \code{character} scalar giving the name of the outcome.}

\item{Yhat}{A \code{character} scalar giving the name of the constructed
predictor.}

\item{x0, x1}{Scalar values giving the two levels of the binary protected
attribute.}

\item{ylvl}{A value indicating within which group the decomposition should be
performed. For example, setting \code{ylvl = 1} would correspond to
performing causal effect decompositions for individuals who have
true outcome \code{Y == 1}.}

\item{method}{A \code{character} scalar with two options: \code{"medDML"} for
mediation double-machine learning, \code{"causal_forest"} for the
\code{\link[grf:causal_forest]{grf::causal_forest()}} method from the \code{grf} package.}

\item{model}{A \code{character} scalar taking values in
\code{c("ranger", "linear")}, indicating whether a tree-based learner is used
\code{\link[ranger:ranger]{ranger::ranger()}}, or if the fitted model should be linear. This parameter
is only relevant if \code{method == "medDML"}.}

\item{tune_params}{A \code{logical(1L)} indicating whether the parameters
should be tuned for the tree-based methods (only the \code{min.node.size}
parameter is tuned). Defaults to \code{FALSE}.}

\item{nboot1}{An \code{integer} scalar determining the number of outer
bootstrap repetitions, that is, how many times the fitting procedure is
repeated. Default is \code{1L}.}

\item{nboot2}{An \code{integer} scalar determining the number of inner
bootstrap repetitions, that is, how many bootstrap samples are taken after
the potential outcomes are obtained from the estimation procedure.
Default is \code{100L}.}

\item{...}{Further arguments passed to downstream model fitting functions.}
}
\description{
Implementation of Fairness Cookbook in Causal Fairness Analysis
(Plecko & Bareinboim 2022), for the case when equality of odds is the
measure of interest.
}
\details{
The procedure takes the data as an input, together with
the causal graph given by the Standard Fairness Model, and outputs a causal
decomposition of the EO measure into direct, indirect, and spurious effects.
}
